requirements.txt
openai-whisper
moviepy


# pip install -r requirements.txt
